+++ src/main/java/com/winwin/winwin/payload/OrganizationRequestPayload.java
@@ -1,5 +1,7 @@
 package com.winwin.winwin.payload;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -23,8 +25,8 @@ public class OrganizationRequestPayload {
 	String name;
 	String ein;
 	String notes;
-	Long revenue;
-	Long assets;
+	BigDecimal revenue;
+	BigDecimal assets;
 	String sector;
 	String sectorLevel;
 	String sectorLevelName;
@@ -37,7 +39,7 @@ public class OrganizationRequestPayload {
 	Boolean isActive;
 	Boolean isTaggingReady;
 	String tagStatus;
-	Long totalAssets;
+	BigDecimal totalAssets;
 	String websiteUrl;
 	String facebookUrl;
 	String linkedinUrl;
